Create game stat helper class, spec file
