+++ spec/stat_tracker_spec.rb
@@ -35,44 +35,8 @@ RSpec.describe StatTracker do
   end
 
   context "Game Statistics" do
-    it "#highest_total_score" do
-      expect(@stat_tracker.highest_total_score).to eq(11)
-    end
-
-    it "#lowest_total_score" do
-      expect(@stat_tracker.lowest_total_score).to eq(0)
-    end
-
     it "#helper total_games" do
       expect(@stat_tracker.total_games).to eq 7441
-    end
-
-    it "#percentage_home_wins" do
-      expect(@stat_tracker.percentage_home_wins).to eq 0.44
-    end
-
-    it "#percentage_visitor_wins" do
-      expect(@stat_tracker.percentage_visitor_wins).to eq 0.36
-    end
-
-    it "#percentage_ties" do
-      expect(@stat_tracker.percentage_ties).to eq 0.20
-    end
-
-    it "#count_of_games_by_season" do
-    expected = {
-        "20122013"=>806,
-        "20162017"=>1317,
-        "20142015"=>1319,
-        "20152016"=>1321,
-        "20132014"=>1323,
-        "20172018"=>1355
-      }
-      expect(@stat_tracker.count_of_games_by_season).to eq(expected)
-    end
-
-    it "#average_goals_per_game" do
-      expect(@stat_tracker.average_goals_per_game).to eq 4.22
     end
 
     it "#helper total_goals_per_season" do
@@ -85,18 +49,6 @@ RSpec.describe StatTracker do
         "20172018"=>6019.0
       }
       expect(@stat_tracker.total_goals_per_season).to eq(expected)
-    end
-
-    it "#average_goals_by_season" do
-      expected = {
-        "20122013"=>4.12,
-        "20162017"=>4.23,
-        "20142015"=>4.14,
-        "20152016"=>4.16,
-        "20132014"=>4.19,
-        "20172018"=>4.44
-      }
-      expect(@stat_tracker.average_goals_by_season).to eq expected
     end
   end
